[Silver II] Title: 이동하기, Time: 432 ms, Memory: 74752 KB -BaekjoonHub

+++ 백준/Silver/11048. 이동하기/이동하기.java
@@ -9,12 +9,12 @@ public class Main {
         StringTokenizer st = new StringTokenizer(br.readLine());
         int n = Integer.parseInt(st.nextToken());
         int m = Integer.parseInt(st.nextToken());
-        int[][] A = new int[n + 1][m + 1];
+        byte[][] A = new byte[n + 1][m + 1];
         int[][] DP = new int[n + 1][m + 1]; // DP배열을 채울 때 경계부분을 따로 처리하지 않기 위해 margin을 한 칸씩 둠.
         for (int i = 1; i <= n; i++) {
             st = new StringTokenizer(br.readLine());
             for (int j = 1; j <= m; j++) {
-                A[i][j] = Integer.parseInt(st.nextToken());
+                A[i][j] = Byte.parseByte(st.nextToken());
             }
         }
